split tests to unit and isolated
- modifying sys.modules for tests requires isolated mode
- isolated mode creates problems with encoding modules for unit tests
